Extract NewSessionModal into standalone Svelte component
Moves all 10 modal state vars, loadExistingSessions, browsePath, and
the full {#if showModal} template block out of +page.svelte into
src/lib/components/NewSessionModal.svelte. createSession() in +page
now receives form data via callback instead of reading module-level
state. Modal-specific CSS migrated to component <style>; shared
backdrop/modal base rules kept in +page.svelte for context picker and
plugin confirm modals.
